Basic character movement
Character can move on the Z-axis with vertical input and on the X-axis with horizontal input. Camera needs to be put at an angle in order to see the ground. Also added rotation for the model toward movement direction, but the testing model I used is turned 90 degrees on the Y-axis so they're facing 90 degrees into the wrong direction.