More clearly defined receiver/sender EventAgents
Added list to progression holding IDs that progressed in eventmanager. Also added support for NPCs to be in the middle of a progression chain (eg. NPC will only let you open the door if you talk to another NPC first)
This was a necessity for the progression system going forward, and was necessary for the kind of progression we'll be working with using companion creatures as "keys" for progression in dialogue with NPCs